+++ multi_frame.py
@@ -43,7 +43,7 @@ for i in range(nfiles):
                           qe=0.9,
                           cr_rate=1.
                           )
-    fits.writeto(Path(path, '{}{}.fits'.format(file_name, i)), sim_im)
+    fits.writeto(Path(path, '{}{}.fits'.format(file_name, i)), sim_im.astype(int))
     if plot_images:
         imagesc(sim_im, 'Output Image')
 
@@ -63,7 +63,7 @@ for i in range(nfiles):
                           qe=0.9,
                           cr_rate=0.
                           )
-    fits.writeto(Path(path, '{}{}.fits'.format(file_name, i)), sim_im)
+    fits.writeto(Path(path, '{}{}.fits'.format(file_name, i)), sim_im.astype(int))
     if plot_images:
         imagesc(sim_im, 'Output Image')
